+++ approvaltests/src/main/java/org/approvaltests/reporters/DiffReporter.java
@@ -8,6 +8,6 @@ public class DiffReporter extends FirstWorkingReporter
   public static final DiffReporter INSTANCE = new DiffReporter();
   public DiffReporter()
   {
-    super(WindowsDiffReporter.INSTANCE, MacDiffReporter.INSTANCE, JunitReporter.INSTANCE, QuietReporter.INSTANCE);
+    super(WindowsDiffReporter.INSTANCE, MacDiffReporter.INSTANCE, JunitReporter.INSTANCE, Junit5Reporter.INSTANCE, QuietReporter.INSTANCE);
   }
 }
