+++ approvaltests/src/main/java/org/approvaltests/reporters/JunitReporter.java
@@ -1,35 +1,20 @@
 package org.approvaltests.reporters;
 
-import java.io.File;
-
-import com.spun.util.ObjectUtils;
-import com.spun.util.io.FileUtils;
-
 import junit.framework.TestCase;
 
-public class JunitReporter implements EnvironmentAwareReporter
+public class JunitReporter extends GenericJUnitReporter
 {
   public static final JunitReporter INSTANCE = new JunitReporter();
+
   @Override
-  public void report(String received, String approved) throws Exception
+  void assertEquals(String aText, String rText)
   {
-    String aText = new File(approved).exists() ? FileUtils.readFile(approved) : "";
-    String rText = FileUtils.readFile(received);
-    String approveCommand = "To approve run : " + ClipboardReporter.getAcceptApprovalText(received, approved);
-    System.out.println(approveCommand);
     TestCase.assertEquals(aText, rText);
   }
+
   @Override
-  public boolean isWorkingInThisEnvironment(String forFile)
+  String getClassNameToTestJunitVersion()
   {
-    try
-    {
-      ObjectUtils.loadClass("junit.framework.TestCase");
-    }
-    catch (Throwable t)
-    {
-      return false;
-    }
-    return GenericDiffReporter.isFileExtensionValid(forFile, GenericDiffReporter.TEXT_FILE_EXTENSIONS);
+    return "junit.framework.TestCase";
   }
 }
